Add files via upload
This .qmd file explores real-life data on Anolis cristatellus from Puerto Rico. It contains an analysis on body temperature vs ambient temperature showing how to model this relation as a least squares linear regression, briefly explaining the least squares principle to draw the best fitting model with the least amount of squared error. At the end it has a task on comparing body temperatures of populations of A. cristatellus in natural settings vs. individuals in urban areas